\name{CompoundDb.Hsapiens.HMDB.4.0}
\docType{package}

\alias{CompoundDb.Hsapiens.HMDB.4.0-package}
\alias{CompoundDb.Hsapiens.HMDB.4.0}
\alias{CompoundDb.Hsapiens.HMDB.4.0}


\title{HMDB compound annotation package}

\description{
  This package contains a \code{\link[PeakABro]{CompoundDb}} database
  providing compound annotations based on HMDB (version 4.0).
  For examples and help on functions see the help pages from the
  \code{PeakABro} package.
}

\author{Johannes Rainer}

\examples{
## load the library
library(CompoundDb.Hsapiens.HMDB.4.0)
## list the contents that are loaded into memory
ls('package:CompoundDb.Hsapiens.HMDB.4.0')
## show the db object that is loaded by calling it's name
CompoundDb.Hsapiens.HMDB.4.0

## for more examples see the PeakABro package.


}

\keyword{package}
\keyword{data}
